playground/merchant-jokes/pyproject.toml
[project]
name = "merchant-jokes"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.135.1",
  "uvicorn>=0.41.0",
  "sangria-merchant-sdk",
  "python-dotenv>=1.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv.sources]
sangria-merchant-sdk = { path = "../../sdk/python", editable = true }
